Move output to a seperate package and remove old output in internal/commands
