+++ internal/commands/load_plugins.go
@@ -3,7 +3,7 @@ package commands
 import (
 	"context"
 	"fmt"
-	
+
 	"github.com/open-policy-agent/conftest/plugin"
 	"github.com/spf13/cobra"
 )
