+++ output/output.go
@@ -7,7 +7,11 @@ import (
 	"io"
 	"log"
 	"os"
+	"runtime"
+	"strings"
 
+	"github.com/jstemmer/go-junit-report/formatter"
+	"github.com/jstemmer/go-junit-report/parser"
 	"github.com/logrusorgru/aurora"
 	table "github.com/olekukonko/tablewriter"
 )
@@ -17,6 +21,7 @@ const (
 	outputJSON  = "json"
 	outputTAP   = "tap"
 	outputTable = "table"
+	outputJUnit = "junit"
 )
 
 // ValidOutputs returns the available output formats for reporting tests
@@ -26,6 +31,7 @@ func ValidOutputs() []string {
 		outputJSON,
 		outputTAP,
 		outputTable,
+		outputJUnit,
 	}
 }
 
@@ -40,6 +46,8 @@ func GetOutputManager(outputFormat string, color bool) OutputManager {
 		return NewDefaultTAPOutputManager()
 	case outputTable:
 		return NewDefaultTableOutputManager()
+	case outputJUnit:
+		return NewDefaultJUnitOutputManager()
 	default:
 		return NewDefaultStandardOutputManager(color)
 	}
@@ -80,6 +88,7 @@ func (s *StandardOutputManager) Put(cr CheckResult) error {
 // Flush writes the contents of the managers buffer to the console
 func (s *StandardOutputManager) Flush() error {
 	var totalFailures int
+	var totalExceptions int
 	var totalWarnings int
 	var totalSuccesses int
 
@@ -120,40 +129,58 @@ func (s *StandardOutputManager) Flush() error {
 			printResults(r, "FAIL", aurora.RedFg)
 		}
 
+		for _, r := range cr.Exceptions {
+			printResults(r, "EXCP", aurora.CyanFg)
+		}
+
 		totalFailures += len(cr.Failures)
+		totalExceptions += len(cr.Exceptions)
 		totalWarnings += len(cr.Warnings)
 		totalSuccesses += len(cr.Successes)
 	}
 
-	totalPolicies := totalFailures + totalWarnings + totalSuccesses
+	totalPolicies := totalFailures + totalExceptions + totalWarnings + totalSuccesses
 
-    var outputColor aurora.Color
-    if totalFailures > 0 {
-        outputColor = aurora.RedFg
-    } else if totalWarnings > 0 {
-        outputColor = aurora.YellowFg
-    } else {
-        outputColor = aurora.GreenFg
-    }
+	var outputColor aurora.Color
+	if totalFailures > 0 {
+		outputColor = aurora.RedFg
+	} else if totalWarnings > 0 {
+		outputColor = aurora.YellowFg
+	} else if totalExceptions > 0 {
+		outputColor = aurora.CyanFg
+	} else {
+		outputColor = aurora.GreenFg
+	}
 
-    var pluralSuffixTests string
-    if totalPolicies != 1 {
-        pluralSuffixTests = "s"
-    }
+	var pluralSuffixTests string
+	if totalPolicies != 1 {
+		pluralSuffixTests = "s"
+	}
 
-    var pluralSuffixWarnings string
-    if totalWarnings != 1 {
-        pluralSuffixWarnings = "s"
-    }
+	var pluralSuffixWarnings string
+	if totalWarnings != 1 {
+		pluralSuffixWarnings = "s"
+	}
 
-    var pluralSuffixFailures string
-    if totalFailures != 1 {
-        pluralSuffixFailures = "s"
-    }
+	var pluralSuffixFailures string
+	if totalFailures != 1 {
+		pluralSuffixFailures = "s"
+	}
 
-    s.logger.Println()
-    outputText := fmt.Sprintf("%v test%s, %v passed, %v warning%s, %v failure%s", totalPolicies, pluralSuffixTests, totalSuccesses, totalWarnings, pluralSuffixWarnings, totalFailures, pluralSuffixFailures)
-    s.logger.Println(s.color.Colorize(outputText, outputColor))
+	var pluralSuffixExceptions string
+	if totalExceptions != 1 {
+		pluralSuffixExceptions = "s"
+	}
+
+	s.logger.Println()
+	outputText := fmt.Sprintf("%v test%s, %v passed, %v warning%s, %v failure%s, %v exception%s",
+		totalPolicies, pluralSuffixTests,
+		totalSuccesses,
+		totalWarnings, pluralSuffixWarnings,
+		totalFailures, pluralSuffixFailures,
+		totalExceptions, pluralSuffixExceptions,
+	)
+	s.logger.Println(s.color.Colorize(outputText, outputColor))
 
 	return nil
 }
@@ -376,4 +403,74 @@ func (t *TableOutputManager) Flush() error {
 	}
 
 	return nil
+}
+
+// JUnitOutputManager formats its output as a JUnit test result
+type JUnitOutputManager struct {
+	p      parser.Package
+	writer io.Writer
+}
+
+// NewDefaultJUnitOutputManager creates a new JUnitOutputManager using standard out
+func NewDefaultJUnitOutputManager() *JUnitOutputManager {
+	return NewJUnitOutputManager(os.Stdout)
+}
+
+// NewJUnitOutputManager creates a new JUnitOutputManager with a given Writer
+func NewJUnitOutputManager(w io.Writer) *JUnitOutputManager {
+	return &JUnitOutputManager{
+		writer: w,
+		p: parser.Package{
+			Name:  "conftest",
+			Tests: []*parser.Test{},
+		},
+	}
+}
+
+// Put puts the result of the check to the manager in the managers buffer
+func (j *JUnitOutputManager) Put(cr CheckResult) error {
+	getOutput := func(r Result) []string {
+		out := []string{
+			r.Message,
+		}
+		for _, err := range r.Traces {
+			out = append(out, err.Error())
+		}
+		return out
+	}
+	convert := func(r Result, status parser.Result) *parser.Test {
+		// We have to make sure that the name of the test is unique
+		name := fmt.Sprintf(
+			"%s - %s",
+			cr.FileName,
+			strings.Split(r.Message, "\n")[0],
+		)
+
+		return &parser.Test{
+			Name:   name,
+			Result: status,
+			Output: getOutput(r),
+		}
+	}
+
+	for _, result := range cr.Warnings {
+		j.p.Tests = append(j.p.Tests, convert(result, parser.FAIL))
+	}
+	for _, result := range cr.Failures {
+		j.p.Tests = append(j.p.Tests, convert(result, parser.FAIL))
+	}
+	for _, result := range cr.Successes {
+		j.p.Tests = append(j.p.Tests, convert(result, parser.PASS))
+	}
+	return nil
+}
+
+// Flush writes the contents of the managers buffer to the console
+func (j *JUnitOutputManager) Flush() error {
+	report := &parser.Report{
+		Packages: []parser.Package{
+			j.p,
+		},
+	}
+	return formatter.JUnitReportXML(report, false, runtime.Version(), j.writer)
 }
