+++ output/result.go
@@ -15,10 +15,11 @@ func (r Result) Error() string {
 // warning and failure "errors" produced by rego should be considered separate
 // from other classes of exceptions.
 type CheckResult struct {
-	FileName  string
-	Warnings  []Result
-	Failures  []Result
-	Successes []Result
+	FileName   string
+	Warnings   []Result
+	Failures   []Result
+	Exceptions []Result
+	Successes  []Result
 }
 
 // NewResult creates a new result from the given message
@@ -31,7 +32,6 @@ func NewResult(message string, traces []error) Result {
 
 	return result
 }
-
 
 func IsResultFailure(result CheckResult, failOnWarn bool) bool {
 	return len(result.Failures) > 0 || (len(result.Warnings) > 0 && failOnWarn)
